Renamed pluginchannel to channel
