+++ src/main/java/net/coreprotect/command/LookupCommand.java
@@ -34,8 +34,8 @@ import net.coreprotect.database.lookup.SignMessageLookup;
 import net.coreprotect.database.statement.UserStatement;
 import net.coreprotect.language.Phrase;
 import net.coreprotect.language.Selector;
-import net.coreprotect.listener.pluginchannel.PluginChannelHandshakeListener;
-import net.coreprotect.listener.pluginchannel.PluginChannelListener;
+import net.coreprotect.listener.channel.PluginChannelHandshakeListener;
+import net.coreprotect.listener.channel.PluginChannelListener;
 import net.coreprotect.utility.Chat;
 import net.coreprotect.utility.ChatMessage;
 import net.coreprotect.utility.Color;
