+++ src/main/java/net/coreprotect/listener/ListenerHandler.java
@@ -15,6 +15,8 @@ import net.coreprotect.listener.block.BlockIgniteListener;
 import net.coreprotect.listener.block.BlockPistonListener;
 import net.coreprotect.listener.block.BlockPlaceListener;
 import net.coreprotect.listener.block.BlockSpreadListener;
+import net.coreprotect.listener.channel.PluginChannelHandshakeListener;
+import net.coreprotect.listener.channel.PluginChannelListener;
 import net.coreprotect.listener.entity.CreatureSpawnListener;
 import net.coreprotect.listener.entity.EntityBlockFormListener;
 import net.coreprotect.listener.entity.EntityChangeBlockListener;
@@ -46,8 +48,6 @@ import net.coreprotect.listener.player.PlayerQuitListener;
 import net.coreprotect.listener.player.PlayerTakeLecternBookListener;
 import net.coreprotect.listener.player.ProjectileLaunchListener;
 import net.coreprotect.listener.player.SignChangeListener;
-import net.coreprotect.listener.pluginchannel.PluginChannelHandshakeListener;
-import net.coreprotect.listener.pluginchannel.PluginChannelListener;
 import net.coreprotect.listener.world.ChunkPopulateListener;
 import net.coreprotect.listener.world.LeavesDecayListener;
 import net.coreprotect.listener.world.PortalCreateListener;
